pyproject.toml
[project]
name = "avtc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.1.0",
    "faiss-cpu>=1.13.2",
    "isort>=7.0.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.12",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.3.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.2.2",
    "tqdm>=4.67.3",
]
